+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-11-19
+
+### Changed
+
+- Updated backup script to use temporary files for archive and encryption
+- Compatibility with Marina 0.7.x
+
 ## [0.1.1] - 2025-11-16
 
 ### Fixed
@@ -19,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of Marina Encrypted Azure Blob Storage Backend
 
-[Unreleased]: https://github.com/polarfoxDev/marina-backend-encrypted-azure/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/polarfoxDev/marina-backend-encrypted-azure/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/polarfoxDev/marina-backend-encrypted-azure/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/polarfoxDev/marina-backend-encrypted-azure/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/polarfoxDev/marina-backend-encrypted-azure/releases/tag/v0.1.0